pyproject.toml
[project]
name = "gemini-slack-bot"
version = "0.1.0"
description = "Geminiと会話できるSlackボット"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "slack-bolt==1.18.0",
    "google-genai>=1.2.0",
    "python-dotenv==1.0.0",
    "flask>=3.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
